Meaningless Commits Inc presents: Made jASSUME to be more scope-proof, and fixed a minor signed/unsigned warning in cdvd.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@1240 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/CDVD/CDVD.cpp
@@ -812,7 +812,7 @@ static uint cdvdStartSeek( uint newsector, CDVD_MODE_TYPE mode )
 {
 	cdvd.SeekToSector = newsector;
 
-	uint delta = abs(cdvd.SeekToSector - cdvd.Sector);
+	uint delta = abs( (s32)(cdvd.SeekToSector - cdvd.Sector) );
 	uint seektime;
 
 	cdvd.Ready = CDVD_NOTREADY;
